refactor: extract account strategy view builder
